# openapi.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateCategoryRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Glass Panels"
          },
          "description": {
            "type": "string",
            "example": "All types of flat glass products"
          }
        }
      },
      "UpdateCategoryRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "CreateProductRequest": {
        "type": "object",
        "required": [
          "name",
          "unit",
          "costPriceUzs",
          "retailPriceUzs",
          "wholesalePriceUzs"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Float Glass 4mm"
          },
          "description": {
            "type": "string",
            "example": "Clear float glass, thickness 4mm"
          },
          "sku": {
            "type": "string",
            "example": "FG-4MM-CLR"
          },
          "unit": {
            "$ref": "#/components/schemas/ProductUnit"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "branchId": {
            "type": "string",
            "format": "uuid",
            "description": "Branch where an initial zero-balance inventory row is created"
          },
          "costPriceUzs": {
            "type": "number",
            "example": 180000
          },
          "retailPriceUzs": {
            "type": "number",
            "example": 250000
          },
          "wholesalePriceUzs": {
            "type": "number",
            "example": 200000
          },
          "costPriceUsd": {
            "type": "number",
            "example": 14
          },
          "retailPriceUsd": {
            "type": "number",
            "example": 19.99
          },
          "wholesalePriceUsd": {
            "type": "number",
            "example": 16.5
          }
        }
      },
...
    }
  },
  "openapi": "3.0.0",
  "info": {
    "title": "Store Management API",
    "version": "1.0.0",
    "description": "Multi-branch retail management system API"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/uploads/organizations/{storeId}/products/{productId}/{imageId}/{fileName}": {
      "get": {
        "summary": "Read an authenticated product image variant",
        "tags": [
          "Product Images"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "imageId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "fileName",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "main.webp",
                "thumbnail.webp"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "WebP image bytes",
            "content": {
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Image not found in the authenticated store"
          }
        }
      }
    },
    "/products/categories": {
      "post": {
        "summary": "Create a product category",
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created"
          },
          "409": {
            "description": "Category name already exists"
          },
          "422": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "summary": "List all categories",
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of categories"
          }
        }
      }
    },
    "/products/categories/{id}": {
      "get": {
        "summary": "Get a category by ID",
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category found"
          },
          "404": {
            "description": "Category not found"
          }
        }
      },
      "patch": {
        "summary": "Update a category",
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated"
          },
          "404": {
            "description": "Category not found"
          },
          "409": {
            "description": "Name already in use"
          }
        }
      },
      "delete": {
        "summary": "Delete a category (only if no products assigned)",
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted"
          },
          "409": {
            "description": "Category has assigned products"
          }
        }
      }
    },
    "/products": {
      "post": {
        "summary": "Create a new product",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          },
          "409": {
            "description": "SKU already in use"
          },
          "422": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "summary": "List products with optional filters",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search by name, SKU, or description"
          },
          {
            "in": "query",
            "name": "categoryId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "unit",
            "schema": {
              "$ref": "#/components/schemas/ProductUnit"
            }
          },
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of products"
          }
        }
      }
    },
    "/products/{productId}/images": {
      "post": {
        "summary": "Upload up to five optimized product images",
        "tags": [
          "Product Images"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "images"
                ],
                "properties": {
                  "images": {
                    "type": "array",
                    "maxItems": 5,
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Images uploaded"
          },
          "413": {
            "description": "Image exceeds 5 MB"
          },
          "422": {
            "description": "Invalid image or product image limit exceeded"
          }
        }
      },
      "get": {
        "summary": "List product images",
        "tags": [
          "Product Images"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ordered product images"
          }
        }
      }
    },
    "/products/{productId}/images/{imageId}": {
      "put": {
        "summary": "Replace a product image while preserving its order and primary state",
        "tags": [
          "Product Images"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "images"
                ],
                "properties": {
                  "images": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 1,
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a product image and select a new primary when needed",
        "tags": [
          "Product Images"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/products/{productId}/images/reorder": {
      "patch": {
        "summary": "Replace the complete product image order",
        "tags": [
          "Product Images"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderProductImagesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Images reordered"
          }
        }
      }
    },
    "/products/{productId}/images/{imageId}/primary": {
      "patch": {
        "summary": "Set the primary product image",
        "tags": [
          "Product Images"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Primary image updated"
          }
        }
      }
    },
    "/products/sku/{sku}": {
      "get": {
        "summary": "Lookup a product by SKU (barcode scan)",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sku",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "FG-4MM-CLR"
          }
        ],
        "responses": {
          "200": {
            "description": "Product found"
          },
          "404": {
            "description": "No product with that SKU"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get a product by ID",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found"
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "patch": {
        "summary": "Update a product",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated"
          },
          "404": {
            "description": "Product or category not found"
          },
          "409": {
            "description": "SKU already in use"
          }
        }
      },
      "delete": {
        "summary": "Delete a product",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
...
  },
...
}